requirements.txt
python-telegram-bot==20.3
pyrogram==2.0.106
tgcrypto==1.2.5
openai==1.3.6
feedparser==6.0.10
python-dotenv==1.0.0
